pyproject.toml
[project]
name = "llm-wiki-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "requests>=2.33.1",
    "schedule>=1.2.2",
]
